Update reqs to support Django 4 and remove Django 2.2
This removes the top end version restriction on Django and bumps the lowest version to 3.0. There are also updates to a couple of other libraries (Djoser, social-auth) as they were out of date and causing issues, and a small update to the testapp/utils.py as the middleware API has changed slightly in 4.0.