Fix lint and fix operation of Ruff and apply changes
Some of the rules in Ruff need to have the scope of the package in order to determine
whether they apply (e.g. I001 for sorting imports). This increases the default batch
size from 128 files to 1024 to avoid issues with only some files in a package being in
scope during execution.