Updating codecov packages, swapping pytest-lazyfixture with pytest-lazy-fixtures
pytest-lazy-fixtures actually appears to be maintained - the old one is not, and it was generating a bunch of errors.
Specifically this one: `AttributeError: 'CallSpec2' object has no attribute 'funcargs'`
The replacement seems to be more actively developed (the last release was <6 months ago at time of commit) and seems to work in the same way, so considering it a drop-in replacement. The tests pass, anyway.