Handful of changes to fix build issues, migrate to uv
- Update the pyproject.toml for using uv rather than rye
- Added a new target in the Dockerfile for building. A lot of the problems I was having with building stuff came down to permissions - there's no real good way to ensure the uid/gid match when bind mounting and that screws up git a whole lot.
- Fixed file paths in the release command (__init__.py isn't where it thought it was)
- Changed the build_support cli to be runnable (ostensibly via uv but maybe you can just do this from the command line too now)
- Update pyproject.tomls for the subprojects so that bumpver can figure out where the __init__.py it's supposed to futz with is

The docs haven't changed quite yet but this all prepares this repo for management with uv. uv is pretty close to rye but there are some differences (and those differences will get smaller over time). uv is made by the same people that make rye and ruff so this is a natural evoltuion of those tools to an extent.