Worked out some stuff so `uv run build` works now
It's best to run this as part of the `release` compose service - the build-support app does stuff(tm) with Git and it will be unhappy about trying to share a bind mount unless you jump through some hoops with users, etc.
Plus, the `release` target sets the python path correctly for the script.