Wraps the CyberSource imports in a warnings filter, so we get less chatter from SyntaxWarning issues (#191)
The CyberSource Python REST client library uses a couple of regexes that aren't in raw strings, and (starting with Python 3.12) doing that raises a SyntaxWarning. (This was deprecated in 3.6 but in between then and 3.12 it raised a DeprecationWarning, which is usually invisible.) Wrapping these in a `catch_warnings` context manager lets us filter out SyntaxWarning, so we can make it quiet. (pytest also sometimes regards these as _errors_, which keeps your tests from passing when using Python 3.12+, so this should fix that too.) It would be better for the upstream to be fixed but we can do this until that happens.